Fixed some minor MSVC warnings

+++ src/opnmidi_midiplay.cpp
@@ -1606,7 +1606,7 @@ size_t OPNMIDIplay::chooseDevice(const std::string &name)
     size_t n = m_midiDevices.size() * 16;
     m_midiDevices.insert(std::make_pair(name, n));
     m_midiChannels.resize(n + 16);
-    resetMIDIDefaults(n);
+    resetMIDIDefaults(static_cast<int>(n));
     return n;
 }
